requirements.txt
flask
flask-socketio
eventlet
pandas

textblob
transformers
torch

nltk
scikit-learn
scipy
numpy
gensim
spacy

groq
requests
python-dotenv

pytest
pytest-cov